+++ apps/accounts/apps.py
@@ -4,3 +4,6 @@ class AccountsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'apps.accounts'
     verbose_name = '認証と認可(ログインユーザ)'
+    
+    def ready(self):
+        import apps.accounts.signals